backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
livekit==0.8.0
livekit-agents==0.4.0
openai==1.3.0
twilio==8.10.0
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
